backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mutagen
python-multipart
psycopg2-binary
boto3
python-dotenv


